Add type completion and refactor Joi error types. Organise errors to a seperate folder.
