+++ src/interfaces/errors/enums.ts
