+++ src/interfaces/errors/joi.ts
@@ -1,0 +1,13 @@
+// ******************** //
+// Interfaces for the possible Joi error types.
+// ******************** //
+
+export interface JoiErrorTypes {
+    TYPE_REQUIRED: string;
+    TYPE_EMPTY: string;
+    TYPE_MIN: string;
+    TYPE_MAX: string;
+    TYPE_LENGTH: string;
+    TYPE_INVALID_EMAIL_FORMAT: string;
+    TYPE_REGEX: string;
+}
